fix: rename from libraryProfile to publicProfile
