+++ src/api/index.ts
@@ -19,4 +19,4 @@ export * from './membership';
 export * from './mentions';
 export * from './search';
 export * from './subscription';
-export * from './libraryProfile';
+export * from './publicProfile';
