+++ src/api/routes.ts
@@ -396,10 +396,10 @@ export const buildGetEtherpadRoute = (itemId: UUID) =>
 
 export const SEARCH_PUBLISHED_ITEMS_ROUTE = `${ITEMS_ROUTE}/${COLLECTIONS_ROUTE}/search`;
 
-export const LIBRARY_PROFILE_ROUTE = `member-profile`;
-export const GET_OWN_LIBRARY_PROFILE = `${LIBRARY_PROFILE_ROUTE}/own`;
-export const buildGetMemberProfileRoute = (memberId: UUID) =>
-  `${LIBRARY_PROFILE_ROUTE}/${memberId}`;
+export const PUBLIC_PROFILE_ROUTE = `member-profile`;
+export const GET_OWN_PROFILE = `${PUBLIC_PROFILE_ROUTE}/own`;
+export const buildGetPublicProfileRoute = (memberId: UUID) =>
+  `${PUBLIC_PROFILE_ROUTE}/${memberId}`;
 
 export const API_ROUTES = {
   APPS_ROUTE,
@@ -504,7 +504,7 @@ export const API_ROUTES = {
   SIGN_OUT_ROUTE,
   SIGN_UP_ROUTE,
   VALIDATION_ROUTE,
-  LIBRARY_PROFILE_ROUTE,
-  GET_OWN_LIBRARY_PROFILE,
-  buildGetMemberProfileRoute,
+  PUBLIC_PROFILE_ROUTE,
+  GET_OWN_PROFILE,
+  buildGetPublicProfileRoute,
 };
