+++ src/hooks/index.ts
@@ -13,11 +13,11 @@ import configureItemLoginHooks from './itemLogin';
 import configureItemPublishedHooks from './itemPublish';
 import configureItemTagHooks from './itemTag';
 import configureItemValidationHooks from './itemValidation';
-import configureLibraryProfileHooks from './libraryProfile';
 import configureMemberHooks from './member';
 import configureMembershipHooks from './membership';
 import configureMentionsHooks from './mention';
 import configurePlanHooks from './plan';
+import configurePublicProfileHooks from './publicProfile';
 import configureKeywordSearchHooks from './search';
 
 export default (
@@ -52,6 +52,6 @@ export default (
     ...configureInvitationHooks(queryConfig),
     ...memberHooks,
     ...configurePlanHooks(queryConfig),
-    ...configureLibraryProfileHooks(queryConfig),
+    ...configurePublicProfileHooks(queryConfig),
   };
 };
